pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "gamma-rader"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2.36",
    "pykrx>=1.0.46",
    "google-generativeai>=0.8.0",
    "google-genai>=1.0.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "feedparser>=6.0.10",
    "pytz>=2024.1",
    "pyyaml>=6.0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
pythonpath = ["."]
